backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
requests
SQLAlchemy
python-dotenv
psycopg2-binary
google-generativeai